+++ src/dda/config/model/telemetry.py
@@ -1,0 +1,21 @@
+# SPDX-FileCopyrightText: 2025-present Datadog, Inc. <dev@datadoghq.com>
+#
+# SPDX-License-Identifier: MIT
+from __future__ import annotations
+
+from typing import Literal
+
+from msgspec import Struct
+
+
+class TelemetryConfig(Struct, frozen=True):
+    """
+    /// tab | :octicons-file-code-16: config.toml
+    ```toml
+    [telemetry]
+    id = "anon"
+    ```
+    ///
+    """
+
+    id: Literal["anon", "user"] = "anon"