+++ tests/cli/self/telemetry/test_id.py
@@ -1,0 +1,16 @@
+# SPDX-FileCopyrightText: 2025-present Datadog, Inc. <dev@datadoghq.com>
+#
+# SPDX-License-Identifier: MIT
+from __future__ import annotations
+
+
+def test(dda, helpers, machine_id):
+    result = dda("self", "telemetry", "id")
+    result.check(
+        exit_code=0,
+        stdout=helpers.dedent(
+            f"""
+            {machine_id}
+            """
+        ),
+    )